ru.cfg
Made Russian Translation - tested and working. Added some small flavor with jokes and references to pop-culture.
Some of mod additions are not translated in the game because there is no lines in localization files, for ex. Drill description additions "...It can only extract Ore deposits of concentrations in excess of 2.5% and Gold Ore/Unobtainium deposits in excess of 2%. Scientists extract more Unobtainium."

Feel free to ask if you have any questions regarding translation